Move soils and environment from simulation file to separate files
