+++ include/pops/model.hpp
@@ -24,6 +24,7 @@
 #ifndef POPS_MODEL_HPP
 #define POPS_MODEL_HPP
 
+#include "environment.hpp"
 #include "config.hpp"
 #include "treatments.hpp"
 #include "spread_rate.hpp"
@@ -32,6 +33,7 @@
 #include "kernel.hpp"
 #include "scheduling.hpp"
 #include "quarantine.hpp"
+#include "soils.hpp"
 
 #include <vector>
