+++ include/pops/simulation.hpp
@@ -28,20 +28,9 @@
 #include <stdexcept>
 
 #include "utils.hpp"
+#include "soils.hpp"
 
 namespace pops {
-
-/** Rotate elements in a container to the left by one
- *
- * Rotates (moves) elements in a container to the left (anticlockwise)
- * by one. The second element is moved to the front and the first
- * element is moved to the back.
- */
-template<typename Container>
-void rotate_left_by_one(Container& container)
-{
-    std::rotate(container.begin(), container.begin() + 1, container.end());
-}
 
 /** Draws n elements from a vector. Expects n to be equal or less than v.size().
  */
@@ -120,120 +109,6 @@ inline ModelType model_type_from_string(const char* text)
     // call the string version
     return model_type_from_string(text ? std::string(text) : std::string());
 }
-
-template<typename IntegerRaster, typename FloatRaster, typename RasterIndex = int>
-class Environment
-{
-public:
-    Environment() {}
-
-    void update_weather_coefficient(const FloatRaster& raster)
-    {
-        current_weather_coefficient = &raster;
-    }
-
-    double weather_coefficient_at(RasterIndex row, RasterIndex col) const
-    {
-        if (!current_weather_coefficient) {
-            throw std::logic_error("Weather coefficient used, but not provided");
-        }
-        return current_weather_coefficient->operator()(row, col);
-    }
-
-protected:
-    const FloatRaster* current_weather_coefficient{nullptr};
-};
-
-template<typename IntegerRaster, typename FloatRaster, typename RasterIndex = int>
-class SoilPool
-{
-public:
-    SoilPool(
-        std::vector<IntegerRaster>& rasters,
-        const Environment<IntegerRaster, FloatRaster, RasterIndex>& environment,
-        bool generate_stochasticity = true,
-        bool establishment_stochasticity = true,
-        double fixed_soil_probability = 0)
-        : rasters_(&rasters),
-          environment_(&environment),
-          generate_stochasticity_(generate_stochasticity),
-          establishment_stochasticity_(establishment_stochasticity),
-          fixed_soil_probability_(fixed_soil_probability)
-    {
-        if (rasters.empty()) {
-            throw std::logic_error(
-                "List of rasters of SpoilPool needs to have at least one item");
-        }
-    }
-
-    template<typename Generator>
-    int dispersers_from(RasterIndex row, RasterIndex col, Generator& generator)
-    {
-        auto count = this->total_at(row, col);
-        double lambda = environment_->weather_coefficient_at(row, col);
-        if (this->generate_stochasticity_) {
-            std::poisson_distribution<int> distribution(lambda);
-            int dispersers = 0;
-            for (int k = 0; k < count; k++) {
-                dispersers += distribution(generator);
-            }
-            return dispersers;
-        }
-        else {
-            return lambda * count;
-        }
-        // TODO: subtract as with mortality moves
-    }
-
-    template<typename Generator>
-    void disperser_to(RasterIndex row, RasterIndex col, Generator& generator)
-    {
-        double current_probability = environment_->weather_coefficient_at(row, col);
-        double tester = 1 - fixed_soil_probability_;
-        if (this->establishment_stochasticity_)
-            tester = distribution_uniform_(generator);
-        if (tester < current_probability) {
-            this->add_at(row, col);
-        }
-    }
-
-    template<typename Generator>
-    void dispersers_to(
-        int dispersers, RasterIndex row, RasterIndex col, Generator& generator)
-    {
-        for (int i = 0; i < dispersers; i++)
-            this->disperser_to(row, col, generator);
-    }
-
-    void add_at(RasterIndex row, RasterIndex col, int value = 1)
-    {
-        rasters_->back()(row, col) += value;
-    }
-
-    int total_at(RasterIndex row, RasterIndex col)
-    {
-        int total = 0;
-        for (auto& raster : *rasters_) {
-            total += raster(row, col);
-        }
-        return total;
-    }
-
-    void next_step(int step)
-    {
-        UNUSED(step);
-        rotate_left_by_one(*rasters_);
-        rasters_->back().fill(0);
-    }
-
-protected:
-    std::vector<IntegerRaster>* rasters_{nullptr};
-    const Environment<IntegerRaster, FloatRaster, RasterIndex>* environment_{nullptr};
-    bool generate_stochasticity_{false};
-    bool establishment_stochasticity_{false};
-    double fixed_soil_probability_{0};
-    std::uniform_real_distribution<double> distribution_uniform_{0.0, 1.0};
-};
 
 /*! The main class to control the spread simulation.
  *
